TH-17 Create image upload component
