+++ client/components/Organisms/EditOrCreatePersonForm/styledComponents.ts
@@ -1,5 +1,5 @@
 import { styled } from "@mui/material/styles";
-import { Box, IconButton, Typography } from "@mui/material";
+import { Box, Typography } from "@mui/material";
 
 export const Container = styled(Box)(() => ({
   width: "700px",
@@ -9,30 +9,6 @@ export const Container = styled(Box)(() => ({
   flexDirection: "column",
   gap: "40px",
   alignItems: "center",
-}));
-
-export const ImageContainer = styled(Box)(() => ({
-  position: "relative",
-  transform: "scale(1.5)",
-  cursor: "pointer",
-}));
-
-export const StyledFileInput = styled("input")(() => ({
-  display: "none",
-}));
-
-export const StyledIconButton = styled(IconButton)(({ theme }) => ({
-  position: "absolute",
-  top: "38px",
-  right: "38px",
-  width: "35px",
-  height: "35px",
-  background: theme.palette.primary.main,
-  color: theme.palette.primary.contrastText,
-  "& svg": {
-    width: "23px",
-    height: "23px",
-  },
 }));
 
 export const ErrorContainer = styled(Box)(() => ({
